# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * EchoPulse Voice Assistant API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface CommandInput {
  /** The transcribed voice command text */
  query: string;
}

export interface CommandResponse {
  /** Either 'local_router' or 'gemini_api' */
  source: string;
  /** The assistant's response text */
  reply: string;
  query: string;
  /**
     * The regex pattern that matched, if local
     * @nullable
     */
  matchedPattern?: string | null;
  /**
     * A machine-readable action type (e.g. 'task_added', 'time', 'play_music')
     * @nullable
     */
  action?: string | null;
}

export interface Task {
  id: number;
  text: string;
  completed: boolean;
  createdAt: string;
  updatedAt?: string;
}

export interface TaskInput {
  /** @minLength 1 */
  text: string;
}

export interface TaskUpdate {
  /** @minLength 1 */
  text?: string;
  completed?: boolean;
}

export interface CommandLog {
  id: number;
  query: string;
  source: string;
  reply: string;
  /** @nullable */
  matchedPattern?: string | null;
  /** @nullable */
  action?: string | null;
  createdAt: string;
}

export interface Preference {
  id: number;
  category: string;
  platform: string;
  count: number;
  updatedAt: string;
}

export type AssistantStatsTopCategoriesItem = {
  category: string;
  count: number;
};

export interface AssistantStats {
  totalCommands: number;
  localCommands: number;
  geminiCommands: number;
  tasksTotal: number;
  tasksCompleted: number;
  topCategories?: AssistantStatsTopCategoriesItem[];
}

export interface ApiError {
  error: string;
}

export type ListCommandLogsParams = {
limit?: number;
};