# lib/api-zod/src/generated/types/assistantStats.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * EchoPulse Voice Assistant API
 * OpenAPI spec version: 0.1.0
 */
import type { AssistantStatsTopCategoriesItem } from './assistantStatsTopCategoriesItem';

export interface AssistantStats {
  totalCommands: number;
  localCommands: number;
  geminiCommands: number;
  tasksTotal: number;
  tasksCompleted: number;
  topCategories?: AssistantStatsTopCategoriesItem[];
}