# lib/api-zod/src/generated/types/commandInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * EchoPulse Voice Assistant API
 * OpenAPI spec version: 0.1.0
 */

export interface CommandInput {
  /** The transcribed voice command text */
  query: string;
}