# lib/api-zod/src/generated/types/commandLog.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * EchoPulse Voice Assistant API
 * OpenAPI spec version: 0.1.0
 */

export interface CommandLog {
  id: number;
  query: string;
  source: string;
  reply: string;
  /** @nullable */
  matchedPattern?: string | null;
  /** @nullable */
  action?: string | null;
  createdAt: Date;
}