# lib/api-zod/src/generated/types/commandResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * EchoPulse Voice Assistant API
 * OpenAPI spec version: 0.1.0
 */

export interface CommandResponse {
  /** Either 'local_router' or 'gemini_api' */
  source: string;
  /** The assistant's response text */
  reply: string;
  query: string;
  /**
     * The regex pattern that matched, if local
     * @nullable
     */
  matchedPattern?: string | null;
  /**
     * A machine-readable action type (e.g. 'task_added', 'time', 'play_music')
     * @nullable
     */
  action?: string | null;
}