# lib/api-zod/src/generated/types/taskUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * EchoPulse Voice Assistant API
 * OpenAPI spec version: 0.1.0
 */

export interface TaskUpdate {
  /** @minLength 1 */
  text?: string;
  completed?: boolean;
}